build: Make eslint plugins peer dependencies
BREAKING CHANGE: This change moves all plugins referenced in this module
from full to peer dependencies and moves all rules from the `goodeggs/`
namespace back to their default rule names. To migrate to this version,
you'll need to install all plugins as dev dependencies and rename any
rules that begin with `goodeggs/` you've overridden in your
project-local configurations to their default rule name as defined by
the plugin.

History lesson here: Historically, these plugins were made full
dependencies and all rules were renamed under the `goodeggs/` namespace
in order to make this plugin a one-dependency install. This was entirely
to get around some ESLint limitations and was well intentioned, but
isn't ideal in practice because as a consumer of the module, it forces
you to figure out new names for all your rules, and as a plugin
developer it makes plugins difficult to work with and makes this module
more difficult to maintain.

Going forward, the recommended way to consume this module while avoiding
the need to maintain all these plugins as dependencies is to install
`@goodeggs/toolkit` and run ESLint via `getk run lint-es`.
`@goodeggs/toolkit` packages all these plugins for you, eliminating the
need for consumers to maintain these plugins as dependencies in their
projects.