chore: split TypeScript into a separate config
Recent changes to `@typescript-eslint/eslint-plugin`, combined with the
set of TypeScript rules we have enabled, mean that the presence of a
`tsconfig.json` file is no longer optional. This change splits the
TypeScript configuration into its own separate config so that vanilla
JavaScript and Flow projects can continue to consume this project
without needing to create a (useless) `tsconfig.json` file (and,
probably, install TypeScript).