feat: disallow implicit JSON file imports
This changeset forces consumers to specify the `.json` extension when
importing JSON files. Because ESLint merges arrays positionally (really
unfortunate, confusing behavior) it's really easy to end up with
unexpected ESLint configuration when specifying arrays of extensions
like we're doing here. This change makes it unnecessary to override most
`eslint-plugin-import` settings, allowing us to simplify our config and
avoid this footgun.

We specify the `.json` extension in most places already, and adding it
is a very quick fix in the places where we don't, so this seems like an
unobtrusive change.