chore: update plugins, typescript, prettier deps
This changeset updates all `eslint-plugin-` dependencies and any of
their peers (i.e. `prettier` and `typescript`).

This enables a couple new rules, and some of the plugins themselves had
breaking changes (a couple rule renames, but mostly changes to their
recommended rulesets) so this is a breaking change.

This changeset also loosens the peer dependency version around
`typescript` to exactly match the version required by
`@typescript-eslint/parser`. It does the same thing with `prettier` to
match the peer dependency specified by `eslint-plugin-prettier`.