fix: re-add accidentally removed camelcase rule to recommended config
In https://github.com/goodeggs/eslint-plugin-goodeggs/commit/40923b5, we accidentally
removed the `camelcase` rule from the `recommended` config. This was obscured by the fact
that we have a separate TS-specific camelcase rule enabled for TS files and we have a
custom-configured `camelcase` enabled in the `ops` config.

Note: this rule is already disabled in the `typescript` config to avoid having redundant
rules.