ci: update to Node 12 to fix dependency error
Ideally we'd test on Node 8, but it's approaching the end of its LTS
lifetime and we're starting to see development-time errors caused by
packages that no longer support it.

For now, the built plugin should actually work on Node 8, but we should
start upgrading all our apps to Node 12 anyway.