feat: ignore underscore-prefixed variables
Without this it's impossible to range over e.g. a `range()` iterator:

```ts
for (const c of range(0, 10)) {
  // ...
}
```

While I was here, I also noticed that `argsIgnorePattern` allows you to
specify `_` as the full argument name, which would conflict with the
canonical lodash import name (`_`). You must now specify at least one
character after the underscore in order for the variable to be ignored.