fix: mark TypeScript as an optional peer dep
This plugin can be used in projects that don't use TypeScript (i.e.
plain JS or Flow projects). This change tells newer Yarn versions to not
log a peer dependency warning when TypeScript is missing (and as soon as
someone tries to lint using the TypeScript config, they'll get a runtime
error).