fix: actually allow unused vars with _ prefix
We've intended to configure `no-unused-vars` and `@typescript-eslint/no-unused-vars` to allow unused variables with a `_` prefix. However, the rule has a separate pattern option for "variables" (excludes arguments for some reason) and "arguments". We were only actually configuring the args option. The comment above it even stated that it ended to enable it for both.

Example of what currently isn't allowed but should be:
```js
const [_unusedFirst, ...theGoodStuff] = arr;
const {...eww: _unusedKey} = obj;
```