fix: disable all core eslint rules supersed by @typescript-eslint rules
See
https://github.com/typescript-eslint/typescript-eslint/blob/885780d4a2b07e418256b7323d76b18453c14a50/packages/eslint-plugin/README.md#extension-rules.

When a @typescript-eslint "extension" rule is enabled, the corresponding, superseded core
eslint rule must be disabled. We had intended to do this, but it got out of sync because:
- we upgraded @typescript-eslint/eslint-plugin by two major versions in
  https://github.com/goodeggs/eslint-plugin-goodeggs/pull/462 and missed a few of these
- it's _really_ hard and painstaking to keep this in sync, since the rules that are
  enabled by `plugin:@typescript-eslint/recommended` automatically disable the corresponding
  core eslint rules, but when we opt in to additional rules we must disable the
  corresponding core rules ourselves

This commit should not contain any breaking changes. It's only fixing a misconfiguration,
which should prevent misreported errors.

It appears that the reason this wasn't caught before is that previous upgraders disabled
the problematic core rules individually in their repos:
- https://github.com/goodeggs/garbanzo/blob/d9c09dac5079a3b85fc885858bd3323d215feb75/.eslintrc.js#L161
- https://github.com/goodeggs/marketplace-mobile-app/blob/8873eadc141157e3b5589819acc318c09bb666db/.eslintrc.js#L88