chore(deps): upgrade to eslint@8 and others
See https://eslint.org/docs/8.0.0/user-guide/migrating-to-8.0.0.

The most notable changes for us are the addition of these 4 rules to the recommended set
(which we leverage):
- `no-loss-of-precision`
- `no-nonoctal-decimal-escape`
- `no-unsafe-optional-chaining`
- `no-useless-backreference`

These should be pretty rare in Good Eggs codebases.

This required an upgrade from the deprecated babel-eslint to @babel/eslint-parser. See
https://github.com/babel/babel-eslint#note-babel-eslint-is-now-babeleslint-parser-and-has-moved-into-the-babel-monorepo.

This also required upgrading eslint-plugin-unicorn from v12 to v38. See
https://github.com/sindresorhus/eslint-plugin-unicorn/releases.

This also required upgrading eslint-plugin-jest, which involved a few rule renames. See
https://github.com/jest-community/eslint-plugin-jest/blob/main/CHANGELOG.md.